Use this module inside another module (how?)
How would that be done? I want to use the functionality of this module without running from the command line, but instead with function calls. I also want to be able to point the file path for the source images to a specified destination, and possibly resize images to fit within the resulting image. Thanks.